requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.59.0
django-cors-headers==4.3.0
django-filter==23.4
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.13.1
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
pipenv==2023.11.15
pippip==0.0.10
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
